HW3/requirements.txt
google-cloud-storage
functions-framework==3.*
google-cloud-pubsub